// packages/web/src/game/tiles.gen.ts
// Generated by scripts/gen-assets.mjs — do not edit.
export const TILE = {
  "floorA": 0,
  "floorB": 1,
  "wallTop": 2,
  "wallFace": 3,
  "window": 4,
  "rug": 5,
  "rugEdgeN": 6,
  "rugEdgeS": 7,
  "rugEdgeW": 8,
  "rugEdgeE": 9,
  "deskL": 10,
  "deskR": 11,
  "chair": 12,
  "plant": 13,
  "shelfTop": 14,
  "shelfBottom": 15,
  "coffee": 16,
  "whiteboard": 17
} as const;
export const TILE_SIZE = 16;